requirements.txt
langchain_community 
tiktoken  
langchainhub 
chromadb 
langchain 
langgraph  
tavily-python
python-dotenv
google-generativeai
langchain-google-genai
langchain-groq
langchain-nomic
langchain-text-splitters
langchain_mistralai
wikipedia
langchain_huggingface
google-search-results
faiss-cpu
sentence_transformers
youtube_search
streamlit
langchain_experimental